Add Routed Networking Mode
Added new network mode, called "routed", that doesn't rely on the docker0 bridge
but still isolates the network stack from the host. It's based on pure layer 3
routing, basically adding one route in the host and one in the container. The
container has a default route using one end of the veth and the host has a route
to the container IP address via the other end of the veth.